Allow attributed strings
